requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Federated Learning and Privacy
opacus>=1.4.0  # Differential privacy for PyTorch

# Communication and APIs
grpcio>=1.50.0
grpcio-tools>=1.50.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.28.0

# Data and Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
alembic>=1.10.0

# Serialization and Compression
protobuf>=4.21.0
pickle5>=0.0.11
lz4>=4.3.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Monitoring and Logging
structlog>=23.1.0
prometheus-client>=0.16.0

# Development and Testing
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=23.1.0
flake8>=6.0.0
mypy>=1.0.0

# AWS and Deployment
boto3>=1.26.0
docker>=6.0.0